AMLS-4728 | Initial WhichCurriences model split
